requirements.txt
aiosmtplib==5.0.0
amqp==5.3.1
annotated-types==0.7.0
async-timeout==5.0.1
billiard==4.2.4
cachetools==6.2.4
celery==5.3.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect==1.3.0
cssutils==2.9.0
dnspython==2.8.0
email-validator==2.1.0
idna==3.11
Jinja2==3.1.3
kombu==5.6.1
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
premailer==3.10.0
prompt_toolkit==3.0.52
pydantic==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.0.1
requests==2.32.5
six==1.17.0
structlog==24.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
